Fix handling of ZED_WINDOW_{SIZE,POSITION} env vars
