+++ script/start-local-collaboration
@@ -31,7 +31,7 @@ scale_factor=1
 if [[ $resolution_line =~ Retina ]]; then scale_factor=2; fi
 width=$(expr ${screen_size[0]} / 2 / $scale_factor)
 height=${screen_size[1] / $scale_factor}
-y=$(expr $height / 2)
+y=0
 
 position_1=0,${y}
 position_2=${width},${y}
